requirements.test.txt
setuptools
numpy
pylint

pytest
pytest-asyncio==0.21.1 # 0.23 has a bug 2023/12/3
pytest-xdist
PyYAML

# mariadb
